requirements.txt
appdirs
certifi
cffi
charset-normalizer
click
cloudpickle
contourpy
cycler
decorator
docker-pycreds
fonttools
gitdb
GitPython
gym
gym-notices
idna
imageio
imageio-ffmpeg
importlib-metadata
importlib-resources
joblib
kiwisolver
lazy_loader
matplotlib
moviepy
networkx
numpy
opencv-python
opencv-python-headless
packaging
pathtools
Pillow
proglog
protobuf
psutil==5.9.4
pycparser==2.21
pyparsing==3.0.9
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.1
PyRep==4.1.0.3
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.2.1
scipy
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tifffile==2023.2.28
torch==1.11.0
tqdm==4.65.0
typing_extensions==4.4.0
urllib3==1.26.14
wandb
zipp==3.12.0
